fix(manager): fix an error in refreshCell call
If a request from user, ask to exchange some ckb, and the ckb is not enough
for user to exchange, the manager will refresh cells. But if the configuration
 of exchange cell amount is 1, that means the manager will not refresh the cells
since the amount reaches, even though the capacity does not. So this commit is to solve
this problem. To check the cell amount as well as the total capacity.